+++ app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,2 @@
-import { handlers } from "@/app/auth"
+import { handlers } from '@/app/(auth)/auth';
 export const { GET, POST } = handlers;
